+++ src/client/core/modules/ContactForm/contact-form.scss
@@ -1,0 +1,72 @@
+@import '~styles/config/_colors.scss';
+@import '~styles/typography.scss';
+.modal {
+  &__container {
+    &__button {
+      height: 100%;
+      width: 100%;
+      background: transparent;
+      border: transparent;
+      color: $white;
+      font-size: 16px;
+      @include NavLink;
+    }
+  }
+}
+.react-responsive-modal-overlay {
+  background: rgba(0, 0, 0, 0.75);
+  display: flex;
+  align-items: flex-start;
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  overflow-y: auto;
+  overflow-x: hidden;
+  z-index: 1000;
+  padding: 1.2rem;
+}
+
+.react-responsive-modal-modal {
+  max-width: 800px;
+  position: relative;
+  padding: 1.2rem;
+  background: #ffffff;
+  background-clip: padding-box;
+  box-shadow: 0 12px 15px 0 rgba(0, 0, 0, 0.25);
+  margin: 0 auto;
+}
+
+.react-responsive-modal-modalCenter {
+  margin: auto;
+}
+
+.react-responsive-modal-closeButton {
+  position: absolute;
+  top: 14px;
+  right: 14px;
+  border: none;
+  padding: 0;
+  cursor: pointer;
+  background-color: transparent;
+  display: flex;
+}
+
+@keyframes react-responsive-modal-fadeIn {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
+
+@keyframes react-responsive-modal-fadeOut {
+  0% {
+    opacity: 1;
+  }
+  100% {
+    opacity: 0;
+  }
+}
